#220: Add `AloFlux::ofType`
